Break unit tests into two files
